move style sheets
